Replace surface normal calculation with adjacent vertices
I didn't consider the fact that, unlike most 3D shapes, the normals are going to change when rotating a >3D object. For 4D objects at least, it should be possible to store the adjacent vertices on the faces for each vertex, calculate the normal based on that (I really wish I had geometry shaders) then pass that to the fragment shader for lighting.